+++ mlir/include/air/Dialect/AIRRt/AIRRtBase.td
@@ -20,7 +20,6 @@ It enapsulates data structures and functions available in the runtime and
 can be lowered to a combination of standard and LLVM dialects.
   }];
   let useDefaultTypePrinterParser = 1;
-  let useFoldAPI = kEmitFoldAdaptorFolder;
 }
 
 def AIRRt_Event : DialectType<
