+++ mlir/include/air/Util/Outliner.h
@@ -14,6 +14,9 @@
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 
+using namespace mlir;
+using namespace mlir::affine;
+
 namespace xilinx {
 namespace air {
 
@@ -26,7 +29,7 @@ public:
 
   mlir::func::CallOp outline(std::vector<mlir::Operation *> ops,
                              std::string fname = "acap_outline_fn");
-  mlir::func::CallOp outline(mlir::AffineForOp forOp,
+  mlir::func::CallOp outline(AffineForOp forOp,
                              std::string fname = "acap_outline_fn");
 
 private:
